+++ src/Elastic.Documentation.Lambda.LinkIndexUploader/aws-lambda-tools-defaults.json
@@ -8,11 +8,8 @@
   "profile": "",
   "region": "",
   "configuration": "Release",
-  "function-runtime": "provided.al2023",
+  "function-runtime": "provided.al2",
   "function-memory-size": 512,
   "function-timeout": 180,
   "function-handler": "Elastic.Documentation.Lambda.LinkIndexUploader",
-  "msbuild-parameters": "--self-contained true",
-  "package-type": "image",
-  "docker-host-build-output-dir": "./bin/Release/lambda-publish"
 }
